requirements.txt
aniso8601==10.0.0
attrs==25.1.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-JWT-Extended==4.7.1
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
pytz==2024.2
referencing==0.36.2
rpds-py==0.22.3
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3

Flask-Migrate~=4.1.0
marshmallow~=3.26.0
alembic~=1.14.1